yarn e2e-local is now e2e, add new yarn e2e-browserstack
