+++ nightwatch.conf.js
@@ -2,38 +2,49 @@ const {
   ROOT_URL
 } = require('./config');
 
-const nightwatch_config = {
+module.exports = {
   src_folders: './test/e2e/specs/',
   output_folder: './test/e2e/reports',
   page_objects_path: './test/e2e/page_objects',
   globals_path: './test/e2e/globals',
-
-  selenium : {
-    start_process: false,
-    host: 'hub-cloud.browserstack.com',
-    port: 80
+  selenium: {
+    start_process: true,
+    server_path: 'node_modules/selenium-standalone/.selenium/selenium-server/3.5.3-server.jar',
+    log_path: './test/e2e/',
+    host: '127.0.0.1',
+    port: 6666,
+    cli_args: {
+      'webdriver.chrome.driver': 'node_modules/selenium-standalone/.selenium/chromedriver/2.32-x64-chromedriver'
+    }
   },
-
   test_settings: {
     default: {
       launch_url: ROOT_URL,
+      selenium_port: 6666,
+      selenium_host: 'localhost',
+      silent: true,
       desiredCapabilities: {
-        'browserstack.user': process.env.BROWSERSTACK_USER || 'coralproject2',
-        'browserstack.key': process.env.BROWSERSTACK_KEY,
-        'browserstack.local': true,
-        'browserstack.debug': true,
-        'browserstack.networkLogs': true,
-        'browser': 'chrome',
-      }
-    }
-  }
+        browserName: 'chrome',
+        javascriptEnabled: true,
+        acceptSslCerts: true,
+        webStorageEnabled: true,
+        databaseEnabled: true,
+        applicationCacheEnabled: false,
+        nativeEvents: true,
+      },
+      screenshots : {
+        enabled: true,
+        on_failure: true,
+        on_error: true,
+        path: './test/e2e/tests_output'
+      },
+    },
+    headless: {
+      desiredCapabilities: {
+        chromeOptions : {
+          args: ['--headless', '--disable-gpu']
+        },
+      },
+    },
+  },
 };
-
-// Code to copy seleniumhost/port into test settings
-for(const i in nightwatch_config.test_settings){
-  const config = nightwatch_config.test_settings[i];
-  config['selenium_host'] = nightwatch_config.selenium.host;
-  config['selenium_port'] = nightwatch_config.selenium.port;
-}
-
-module.exports = nightwatch_config;
