+++ package.json
@@ -15,9 +15,9 @@
     "lint": "yamllint locales/*.yml && eslint --ext=.js --ext=.json bin/* .",
     "lint-fix": "yarn lint --fix",
     "jest-watch": "TEST_MODE=unit NODE_ENV=test jest --watch",
-    "e2e": "NODE_ENV=test ./scripts/e2e.js",
-    "pree2e-local": "./scripts/e2e-local.sh",
-    "e2e-local": "NODE_ENV=test nightwatch --config nightwatch-local.conf.js",
+    "e2e-browserstack": "NODE_ENV=test ./scripts/e2e.js --config nightwatch-browserstack.conf.js",
+    "pree2e": "selenium-standalone install",
+    "e2e": "NODE_ENV=test nightwatch",
     "test": "TEST_MODE=unit NODE_ENV=test jest && TEST_MODE=unit NODE_ENV=test mocha -R ${MOCHA_REPORTER:-spec}",
     "test-cover": "TEST_MODE=unit NODE_ENV=test istanbul cover _mocha --report text --check-coverage -- -R spec",
     "heroku-postbuild": "./bin/cli plugins reconcile && yarn build",
