--- scripts/e2e-local.sh
@@ -1,1 +1,0 @@
-selenium-standalone install
